{{cookiecutter.project_name}}/djangoapp/requirements.txt
# base
dj-database-url==0.5.0
Django==3.2.4
psycopg2-binary==2.9.1
python-decouple==3.4
pytz==2021.1

# local
django-debug-toolbar==3.2.1
ipython==7.25.0

# production
gunicorn==20.1.0
sentry-sdk==1.1.0

# tests
black==21.6b0
coverage==5.5
factory-boy==3.2.0
flake8==3.9.2
isort==5.9.1
tox==3.23.1